move not_unsafe_ptr_arg_deref to its own module
